Fix test/build failures (win64 compile warning, isolate test parallel compilation failure, gcmole issue, isolate test intermittant failure)
BUG=

Review URL: https://codereview.chromium.org/14265024

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@14449 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/hydrogen.cc
@@ -1860,7 +1860,7 @@ HValue* HGraphBuilder::JSArrayBuilder::EmitMapCode(HValue* context) {
       HGlobalObject(context));
   HInstruction* native_context = AddInstruction(new(zone())
       HLoadNamedField(global_object, true, GlobalObject::kNativeContextOffset));
-  size_t offset = Context::kHeaderSize +
+  int offset = Context::kHeaderSize +
       kPointerSize * Context::JS_ARRAY_MAPS_INDEX;
   HInstruction* map_array = AddInstruction(new(zone())
       HLoadNamedField(native_context, true, offset));
